requirements.txt
mcp[cli]
fastmcp
httpx
asyncssh
websockets
python-dotenv
